build(deps): replace humantime with jiff
Resolves: #110
