+++ Cargo.toml
@@ -14,6 +14,7 @@ ctrlc = "3.2.5"
 figment = { version = "0.10", features = ["toml", "env", "json"] }
 futures = "0.3.29"
 human-bandwidth = { version = "0.1.1", features = ["serde"] }
+jiff = "0.2.4"
 netem-trace = { version = "0.4.0", features = ["serde", "human", "mahimahi"] }
 nix = { version = "0.29.0", features = [
   "time",
@@ -56,7 +57,7 @@ dirs = "6.0.0"
 figment = { workspace = true }
 futures = { workspace = true }
 human-bandwidth = { workspace = true }
-humantime = "2.1.0"
+jiff = { workspace = true }
 netem-trace = { workspace = true }
 nix = { workspace = true }
 once_cell = { workspace = true }
