+++ rattan-core/Cargo.toml
@@ -19,8 +19,8 @@ camellia = { git = "https://github.com/minhuw/camellia.git", optional = true }
 etherparse = "0.17.0"
 futures = { workspace = true }
 human-bandwidth = { workspace = true }
-humantime-serde = { version = "1.1.1", optional = true }
 ipnet = { version = "2.8.0", features = ["serde"] }
+jiff = { workspace = true }
 lazy_static = "1.4.0"
 libc = "0.2.141"
 netem-trace = { workspace = true }
@@ -69,13 +69,7 @@ harness = false
 default = []
 http = ["serde", "dep:axum"]
 packet-dump = ["dep:pcap-file"]
-serde = [
-  "dep:humantime-serde",
-  "dep:serde",
-  "dep:serde_json",
-  "bandwidth/serde",
-  "bytesize/serde",
-]
+serde = ["dep:serde", "dep:serde_json", "bandwidth/serde", "bytesize/serde"]
 camellia = ["dep:camellia"]
 
 [[example]]
